File structure changes to make separate folder out of `EventInfoViews`, for easier searchability/access
